+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/collect_multi_namespace_references.test.mock.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import type { findLegacyUrlAliases } from '../../legacy_url_aliases';
+import type { findLegacyUrlAliases } from './find_legacy_url_aliases';
 import type { findSharedOriginObjects } from '../utils/find_shared_origin_objects';
 import type * as InternalUtils from '../utils/internal_utils';
 
@@ -14,7 +14,7 @@ export const mockFindLegacyUrlAliases = jest.fn() as jest.MockedFunction<
   typeof findLegacyUrlAliases
 >;
 
-jest.mock('../../legacy_url_aliases', () => {
+jest.mock('./find_legacy_url_aliases', () => {
   return { findLegacyUrlAliases: mockFindLegacyUrlAliases };
 });
