+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/delete_legacy_url_aliases.test.mock.ts
@@ -15,6 +15,10 @@ jest.mock('@kbn/core-elasticsearch-client-server-internal', () => {
 });
 
 // Mock this function to return empty results, as this simplifies test cases and we don't need to exercise alternate code paths for these
-jest.mock('../search_dsl', () => {
-  return { getSearchDsl: jest.fn() };
+jest.mock('../../search', () => {
+  const actual = jest.requireActual('../../search');
+  return {
+    ...actual,
+    getSearchDsl: jest.fn(),
+  };
 });
