+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/delete_legacy_url_aliases.ts
@@ -7,7 +7,6 @@
  */
 
 import * as esKuery from '@kbn/es-query';
-
 import { getErrorMessage as getEsErrorMessage } from '@kbn/core-elasticsearch-client-server-internal';
 import type { ISavedObjectTypeRegistry } from '@kbn/core-saved-objects-server';
 import { ALL_NAMESPACES_STRING } from '@kbn/core-saved-objects-utils-server';
@@ -15,8 +14,8 @@ import {
   LEGACY_URL_ALIAS_TYPE,
   type IndexMapping,
 } from '@kbn/core-saved-objects-base-server-internal';
-import type { RepositoryEsClient } from '../repository_es_client';
-import { getSearchDsl } from '../search_dsl';
+import type { RepositoryEsClient } from '../../repository_es_client';
+import { getSearchDsl } from '../../search';
 
 /** @internal */
 export interface DeleteLegacyUrlAliasesParams {
