+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/find_legacy_url_aliases.test.ts
@@ -7,15 +7,14 @@
  */
 
 import { DeeplyMockedKeys } from '@kbn/utility-types-jest';
-
 import {
   type LegacyUrlAlias,
   LEGACY_URL_ALIAS_TYPE,
 } from '@kbn/core-saved-objects-base-server-internal';
-import { CreatePointInTimeFinderFn, PointInTimeFinder } from '../point_in_time_finder';
-import { findLegacyUrlAliases } from './find_legacy_url_aliases';
-import { savedObjectsPointInTimeFinderMock } from '../../mocks';
 import { SavedObjectsPointInTimeFinderClient } from '@kbn/core-saved-objects-api-server';
+import { savedObjectsPointInTimeFinderMock } from '../../../mocks';
+import { CreatePointInTimeFinderFn, PointInTimeFinder } from '../../point_in_time_finder';
+import { findLegacyUrlAliases } from './find_legacy_url_aliases';
 
 describe('findLegacyUrlAliases', () => {
   let pitFinderClientMock: jest.Mocked<SavedObjectsPointInTimeFinderClient>;
