+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/find_legacy_url_aliases.ts
@@ -12,7 +12,7 @@ import {
   LEGACY_URL_ALIAS_TYPE,
   getObjectKey,
 } from '@kbn/core-saved-objects-base-server-internal';
-import type { CreatePointInTimeFinderFn } from '../point_in_time_finder';
+import type { CreatePointInTimeFinderFn } from '../../point_in_time_finder';
 
 interface FindLegacyUrlAliasesObject {
   type: string;
