+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/preflight_check_for_create.test.ts
@@ -12,7 +12,6 @@ import {
 } from './preflight_check_for_create.test.mock';
 
 import type { DeeplyMockedKeys } from '@kbn/utility-types-jest';
-
 import type { ElasticsearchClient } from '@kbn/core-elasticsearch-server';
 import { elasticsearchClientMock } from '@kbn/core-elasticsearch-client-server-mocks';
 import {
