+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/search/aggregations/aggs_types/common_schemas.ts
