+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/search/utils/index.ts
@@ -6,7 +6,10 @@
  * Side Public License, v 1.
  */
 
-export { SavedObjectsRepository } from './repository';
-export { SavedObjectsClientProvider } from './scoped_client_provider';
-export { PointInTimeFinder } from './point_in_time_finder';
-export type { ISavedObjectsClientProvider } from './scoped_client_provider';
+export {
+  fieldDefined,
+  hasFilterKeyError,
+  isSavedObjectAttr,
+  validateConvertFilterToKueryNode,
+  validateFilterKueryNode,
+} from './filter_utils';
