+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/utils/decorate_es_error.ts
